feat: enforce PascalCase for enums
