+++ packages/eslint-plugin-spruce/package.json
@@ -16,5 +16,8 @@
 	"scripts": {},
 	"bugs": {
 		"url": "https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/issues"
+	},
+	"dependencies": {
+		"pascal-case": "^2.0.1"
 	}
 }
